Add writeFile tests
